+++ CHANGELOG.md
@@ -8,8 +8,14 @@ The changes documented here do not include those from the original repository.
 
 ## [Unreleased]
 
+## 2021-09-15
+- Implemented WriteProfileData for Android (https://outsystemsrd.atlassian.net/browse/RMET-976)
+
 ## 2021-09-13
 - Implemeted method to write profile data (https://outsystemsrd.atlassian.net/browse/RMET-995)
+
+## 2021-09-13
+- Fixed RequestPermissions for Android (https://outsystemsrd.atlassian.net/browse/RMET-971)
 
 ## 2021-09-10
 - Added code to check for google play services (https://outsystemsrd.atlassian.net/browse/RMET-830)